переписав функції та змінив мову імен на англійську
