+++ package.json
@@ -27,6 +27,7 @@
   },
   "dependencies": {
     "debounce": "^1.2.1",
+    "mo": "^1.7.3",
     "tui-pagination": "^3.4.1",
     "validator": "^13.6.0"
   }
