+++ src/partials/footer.html
@@ -18,9 +18,9 @@
     ></button>
           <ul class="developers__list">
           <li class="developers__item">
-         <img class="developers__foto" src="./images/icons/ilona.jpg" alt="Ілона Миронюк" width="130" height="150"/>
+         <img class="developers__foto" src="./images/icons/ilona.jpg" alt="Ilona Myronyuk" width="130" height="150"/>
           <div>
-          <h3 class="developers-text">Ілона Миронюк</h3>
+          <h3 class="developers-text">Ilona Myronyuk</h3>
           <ul class="list socials-icon">
             <li class="socials-icon-item">
               <a href="" class="social-icon-link">
@@ -54,9 +54,9 @@
         </div>
         </li>
         <li class="developers__item">
-          <img class="developers__foto" src="./images/icons/artem.jpg" alt="Артем Білодід" width="130" height="150" />
+          <img class="developers__foto" src="./images/icons/artem.jpg" alt="Artem Bilodid" width="130" height="150" />
           <div>
-          <h3 class="developers-text">Артем Білодід</h3>
+          <h3 class="developers-text">Artem Bilodid</h3>
           <ul class="list socials-icon">
             <li class="socials-icon-item">
               <a href="" class="social-icon-link">
@@ -92,7 +92,7 @@
         <li class="developers__item">
           <img class="developers__foto" src="./images/icons/oleksandr.jpg" alt="Олександр Корнієць" width="130" height="150" />
           <div>
-          <h3 class="developers-text">Олександр Корнієць</h3>
+          <h3 class="developers-text">Oleksandr Korniiets</h3>
           <ul class="list socials-icon">
             <li class="socials-icon-item">
               <a href="" class="social-icon-link">
@@ -126,9 +126,9 @@
         </div>
         </li>
         <li class="developers__item">
-          <img class="developers__foto" src="./images/icons/roman.jpg" alt="Роман Ллієнко" width="130" height="150" />
+          <img class="developers__foto" src="./images/icons/roman.jpg" alt="Roman Ilienko" width="130" height="150" />
           <div>
-          <h3 class="developers-text">Роман Ілієнко</h3>
+          <h3 class="developers-text">Roman Ilienko</h3>
           <ul class="list socials-icon">
             <li class="socials-icon-item">
               <a href="" class="social-icon-link">
@@ -162,9 +162,9 @@
         </div>
         </li>
         <li class="developers__item">
-          <img class="developers__foto" src="./images/icons/tanya.jpg" alt="Тетяна Тулук" width="130" height="150" />
+          <img class="developers__foto" src="./images/icons/tanya.jpg" alt="Tatyana Tuluk" width="130" height="150" />
           <div>
-          <h3 class="developers-text">Тетяна Тулук</h3>
+          <h3 class="developers-text">Tatyana Tuluk</h3>
           <ul class="list socials-icon">
             <li class="socials-icon-item">
               <a href="" class="social-icon-link">
@@ -198,9 +198,9 @@
         </div>
         </li>
         <li class="developers__item">
-          <img class="developers__foto" src="./images/icons/vlad.jpg" alt="Владислав Апелганс" width="130" height="150" />
+          <img class="developers__foto" src="./images/icons/vlad.jpg" alt="Vladislav Apelganz" width="130" height="150" />
           <div>
-          <h3 class="developers-text">Владислав Апельганц</h3>
+          <h3 class="developers-text">Vladislav Apelganz</h3>
           <ul class="list socials-icon">
             <li class="socials-icon-item">
               <a href="" class="social-icon-link">
@@ -234,9 +234,9 @@
         </div>
         </li>
         <li class="developers__item">
-          <img class="developers__foto" src="./images/icons/mihail.jpg" alt="Михайло Осіпчук" width="130" height="150" />
+          <img class="developers__foto" src="./images/icons/mihail.jpg" alt="Mikhailo Оsipchuk" width="130" height="150" />
           <div>
-          <h3 class="developers-text">Михайло Осіпчук</h3>
+          <h3 class="developers-text">Mikhailo Оsipchuk</h3>
           <ul class="list socials-icon">
             <li class="socials-icon-item">
               <a href="" class="social-icon-link">
@@ -270,9 +270,9 @@
         </div>
         </li>
         <li class="developers__item">
-          <img class="developers__foto" src="./images/icons/olesay.jpg" alt="Олеся Жук" width="130" height="150" />
+          <img class="developers__foto" src="./images/icons/olesay.jpg" alt="Alesia Zuk" width="130" height="150" />
           <div>
-          <h3 class="developers-text">Олеся Жук</h3>
+          <h3 class="developers-text">Alesia Zuk</h3>
           <ul class="list socials-icon">
             <li class="socials-icon-item">
               <a href="" class="social-icon-link">
